feat: Create EditInformationsProfile fragment and rename files fragment
