+++ app/src/main/java/com/example/ynov_lyon_bde/ui/screens/profil/EditInformationsProfileFragment.kt
@@ -1,0 +1,16 @@
+package com.example.ynov_lyon_bde.ui.screens.profil
+
+import android.os.Bundle
+import androidx.fragment.app.Fragment
+import android.view.LayoutInflater
+import android.view.View
+import android.view.ViewGroup
+import com.example.ynov_lyon_bde.R
+
+class EditInformationsProfileFragment : Fragment() {
+    override fun onCreateView(inflater: LayoutInflater, container: ViewGroup?,
+                              savedInstanceState: Bundle?): View? {
+        // Inflate the layout for this fragment
+        return inflater.inflate(R.layout.fragment_edit_informations_profile, container, false)
+    }
+}
